add utlis filterByPromotion
