+++ src/utlis/filterByPromotion.ts
@@ -1,0 +1,10 @@
+import { BlogPostDTO, IAllowPromotion } from "../type/dto";
+
+const filterByPromotion = (
+  posts: BlogPostDTO[],
+  promotion: IAllowPromotion
+): BlogPostDTO[] => {
+  return posts.filter((post) => post.attributes.promotion === promotion);
+};
+
+export default filterByPromotion;
